ArgopyCarbon split into GreenCoding and Github classes
