+++ argopy/utils/__init__.py
@@ -43,7 +43,7 @@ from .locals import (  # noqa: F401
     get_sys_info,  # noqa: F401
     netcdf_and_hdf5_versions,  # noqa: F401
 )
-from .monitors import monitor_status, badge, fetch_status, ArgopyCarbon  # noqa: F401
+from .monitors import monitor_status, badge, fetch_status  # noqa: F401
 from .geo import (
     wmo2box,
     wrap_longitude,
@@ -61,6 +61,7 @@ from .transform import (
 )
 from .format import argo_split_path, format_oneline, UriCName, redact
 from .loggers import warnUnless, log_argopy_callerstack
+from .carbon import GreenCoding, Github
 
 import importlib
 path2assets = importlib.util.find_spec('argopy.static.assets').submodule_search_locations[0]
@@ -118,7 +119,6 @@ __all__ = (
     "modified_environ",
     # Monitors
     "monitor_status",
-    "ArgopyCarbon",
     # Geo (space/time data utilities)
     "wmo2box",
     "wrap_longitude",
@@ -142,4 +142,7 @@ __all__ = (
     # Loggers:
     "warnUnless",
     "log_argopy_callerstack",
+    # Carbon
+    "GreenCoding",
+    "Github",
 )
