+++ argopy/utils/monitors.py
@@ -1,13 +1,8 @@
-import json
 import urllib
 import importlib
 import time
 import threading
-import pandas as pd
-import copy
-from typing import Union, List, Dict, Literal
 import logging
-from pathlib import Path
 
 
 log = logging.getLogger("argopy.utils.monitors")
@@ -173,420 +168,3 @@ class monitor_status:
         thread = threading.Thread(target=self.work)
         display(self.text)
         thread.start()
-
-
-class ArgopyCarbon:
-    """argopy carbon footprint helper class
-
-    This class uses the `Green-Coding Solutions API <https://api.green-coding.io>`_ to retrieve CI tests energy consumption data.
-
-    This class also uses the `Github API <https://docs.github.com/en/rest>`_, to get PRs and release information.
-
-    Examples
-    --------
-    .. code-block:: python
-        :caption: ArgopyCarbon API
-
-        ArgopyCarbon().workflows
-        ArgopyCarbon().measurements(branch='master', start_date='2025-01-01')
-        ArgopyCarbon().measurements(branch='385/merge', start_date='2025-01-01')
-        ArgopyCarbon().total_measurements(branches=['master', '385/merge'])
-
-        ArgopyCarbon().releases
-        ArgopyCarbon().lastreleasedate
-        ArgopyCarbon().lastPRs
-        ArgopyCarbon().get_PRtitle(385)
-        ArgopyCarbon().get_PRmerged_since('2025-01-01')
-
-        ArgopyCarbon().footprint_since_last_release()
-
-    .. code-block:: python
-        :caption: Get metrics for another repo
-
-        ac = ArgopyCarbon()
-        ac.repo = 'argopy-status'
-        ac.workflows = [{'ID': '2724029', 'Name': 'API status'}]
-        ac.total_measurements()
-
-    """
-
-    owner = "euroargodev"
-    """Github owner parameter, default to 'euroargodev'"""
-
-    repo = "argopy"
-    """Github repo parameter, default to 'argopy'"""
-
-    workflows = [
-        {"ID": "22344160", "Name": "CI tests"},
-        {"ID": "25052179", "Name": "CI tests upstream"},
-    ]
-    """List of github actions workflow parameters"""
-
-    def __init__(self):
-        from ..stores import httpstore
-
-        self.fs = httpstore(cache=True)
-
-    def measurements(
-        self,
-        branch: str = "master",
-        start_date: Union[str, pd.Timestamp] = None,
-        end_date: Union[str, pd.Timestamp] = None,
-        errors: Literal["raise", "ignore", "silent"] = "ignore",
-    ) -> List[Dict]:
-        """Return measurements for a given branch and date period
-
-        Parameters
-        ----------
-        branch : str, default='master'
-            Name of the branch to retrieve measurements for
-        start_date : str, :class:`pandas.Timestamp`, default=None
-            Measurements starting date, default to 1 year before today
-        end_date : str, :class:`pandas.Timestamp`, default=None
-            Measurements ending date, default to today
-        errors: Literal, default: ``ignore``
-            Define how to handle errors raised during data fetching:
-                - ``ignore`` (default): Do not stop processing, simply issue a debug message in logging console
-                - ``raise``: Raise any error encountered
-                - ``silent``:  Do not stop processing and do not issue log message
-
-        Returns
-        -------
-        List[Dict] :
-            List of workflows name, ID and measurements as :class:`pandas.DataFrame`
-
-        See Also
-        --------
-        :class:`ArgopyCarbon.workflows`
-        """
-        if end_date is None:
-            end_date = pd.to_datetime("now", utc=True)
-        else:
-            end_date = pd.to_datetime(end_date)
-        if start_date is None:
-            start_date = end_date - pd.Timedelta(365, unit="D")
-        else:
-            start_date = pd.to_datetime(start_date)
-
-        long_names = [  # noqa: F841
-            "Energy Value [J]",
-            "Run ID",
-            "Ran At",
-            "Label",
-            "CPU",
-            "Commit Hash",
-            "Duration [ms]",
-            "Platform",
-            "Avg. CPU Utilization",
-            "Workflow",
-            "?",
-            "??",
-            "???",
-            "Grid Intensity [gCOE2/kWh]",
-            "gCO2eq of run []",
-        ]
-
-        names = [
-            "Energy",
-            "runID",
-            "Date",
-            "Workflow_Step",
-            "CPU_type",
-            "CommitHash",
-            "Duration",
-            "platform",
-            "Avg. CPU_utilization",
-            "Workflow",
-            "?",
-            "??",
-            "???",
-            "GridIntensity",
-            "gCO2eq",
-        ]
-
-        results = copy.deepcopy(self.workflows)
-        for irow, workflow in enumerate(self.workflows):
-
-            payload = {
-                "repo": f"{self.owner}/{self.repo}",
-                "branch": branch,
-                "workflow": workflow["ID"],
-                "start_date": start_date.strftime("%Y-%m-%d"),
-                "end_date": end_date.strftime("%Y-%m-%d"),
-            }
-
-            uri = (
-                "https://api.green-coding.io/v1/ci/measurements?"
-                + urllib.parse.urlencode(payload)
-            )
-
-            try:
-                data = self.fs.open_json(uri)
-
-                df = pd.DataFrame(data["data"], columns=names)
-                df["Duration"] = df["Duration"] / 1e6  # seconds
-                df["gCO2eq"] = df["gCO2eq"] / 1e6  # gCO2eq
-                df["Energy"] = df["Energy"] / 1e6  # Joules
-
-                df2 = df[["Workflow_Step", "Energy", "Duration", "gCO2eq"]]
-                df2 = df2.groupby("Workflow_Step").sum()
-
-                results[irow]["measurements"] = df2
-
-            except json.JSONDecodeError:
-                msg = (
-                    "No data returned, probably because the branch '%s' was not found at %s"
-                    % (branch, uri)
-                )
-                if errors == "raise":
-                    from ..errors import DataNotFound
-
-                    raise DataNotFound(msg)
-                elif errors == "ignore":
-                    log.debug(msg)
-
-            except Exception as e:
-                if errors == "raise":
-                    raise
-                elif errors == "ignore":
-                    log.error("Error: {e}")
-
-        return results
-
-    def total_measurements(self, branches: List[str] = ["master"], **kwargs) -> float:
-        """Compute the cumulated measurements of gCO2eq for a list of branches and workflows
-
-        Parameters
-        ----------
-        branches : List[str], default = ['master']
-            List of branches to retrieve measurements for.
-            Note that for a given merged PR number, the branche name is ``<PR>/merged``.
-        **kwargs:
-            Other Parameters are passed to :class:`ArgopyCarbon.measurements`
-
-        Returns
-        -------
-        float
-        """
-        gCO2eq = 0
-        for branch in branches:
-            for wkf in self.measurements(branch=branch, **kwargs):
-                if "measurements" in wkf:
-                    gCO2eq += wkf["measurements"]["gCO2eq"].sum()
-        return gCO2eq
-
-    @property
-    def releases(self) -> pd.DataFrame:
-        """List of published releases
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-        """
-        js = self.fs.open_json(
-            f"https://api.github.com/repos/{self.owner}/{self.repo}/releases"
-        )
-        results = []
-        for rel in js:
-            results.append(
-                {
-                    "tag": rel["tag_name"],
-                    "published": not bool(rel["draft"]),
-                    "date": rel["published_at"],
-                }
-            )
-        df = pd.DataFrame(results)
-        df = df.sort_values("date").reset_index(drop=1)
-        return df
-
-    @property
-    def lastreleasedate(self) -> pd.Timestamp:
-        """Last release publication date
-
-        Returns
-        -------
-        :class:`pandas.Timestamp`
-        """
-        df = self.releases
-        return pd.to_datetime(df["date"].max())
-
-    def get_PRtitle(self, pr_num: int) -> str:
-        """Get the title of a given PR number
-
-        Parameters
-        ----------
-        pr_num : int
-
-        Returns
-        -------
-        str
-        """
-        js = self.fs.open_json(
-            f"https://api.github.com/repos/{self.owner}/{self.repo}/pulls/{pr_num}"
-        )
-        return js["title"]
-
-    def get_PRmerged_since(self, start_date: Union[str, pd.Timestamp]) -> pd.DataFrame:
-        """List all PR merged since a given date
-
-        Parameters
-        ----------
-        start_date : :class:`pandas.Timestamp`
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-        """
-        payload = {
-            "state": "closed",
-            "sort": "created",
-            "per_page": 100,
-            "page": 1,
-            "direction": "desc",
-        }
-        # Note that we don't look for other page results and expect less than 100 PRs merged since start_date !
-        uri = (
-            f"https://api.github.com/repos/{self.owner}/{self.repo}/pulls?"
-            + urllib.parse.urlencode(payload)
-        )
-        js = self.fs.open_json(uri)
-        results = []
-        for j in js:
-            PRmerged = pd.to_datetime(j["merged_at"])
-            if PRmerged and PRmerged > pd.to_datetime(start_date, utc=1):
-                results.append(
-                    {
-                        "ID": j["number"],
-                        "title": j["title"],
-                        "merged": pd.to_datetime(j["merged_at"]),
-                    }
-                )
-        df = pd.DataFrame(results).sort_values("merged").reset_index(drop=1)
-        return df
-
-    @property
-    def lastPRs(self) -> pd.DataFrame:
-        """List all PRs merged since the last release
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-
-        See Also
-        --------
-        :class:`ArgopyCarbon.get_PRmerged_since`, :class:`ArgopyCarbon.lastreleasedate`
-        """
-        return self.get_PRmerged_since(self.lastreleasedate)
-
-    def footprint_since_last_release(
-        self, with_master: bool = True, errors="ignore"
-    ) -> float:
-        """Compute total carbon footprint since the last release
-
-        Parameters
-        ----------
-        with_master : bool, default=True
-            Should we consider also the 'master' branch footprint or not.
-        errors: Literal, default: ``ignore``
-            Define how to handle errors raised during data fetching:
-                - ``raise`` (default): Raise any error encountered
-                - ``ignore``: Do not stop processing, simply issue a debug message in logging console
-                - ``silent``:  Do not stop processing and do not issue log message
-
-        Returns
-        -------
-        float
-            Total carbon footprint in gCO2eq, considering all workflows
-
-        See Also
-        --------
-        :class:`ArgopyCarbon.total_measurements`, :class:`ArgopyCarbon.lastreleasedate`, :class:`ArgopyCarbon.lastPRs`
-        """
-        df = self.lastPRs
-        branches = ["%i/merge" % pr for pr in df["ID"]]
-        if with_master:
-            branches.append("master")
-        return self.total_measurements(
-            branches, start_date=self.lastreleasedate, errors=errors
-        )
-
-    def __repr__(self):
-        summary = ["<ArgopyCarbon>"]
-        summary.append("Last release date: %s" % self.lastreleasedate)
-        summary.append("%i PRs merged since the last release" % len(self.lastPRs))
-        summary.append(
-            "Workflows analysed: %s"
-            % ", ".join(
-                ["%s [%s]" % (wkf["Name"], wkf["ID"]) for wkf in self.workflows]
-            )
-        )
-        summary.append(
-            "gCO2eq since last release (including 'master' branch'): %0.2f"
-            % self.footprint_since_last_release(errors="ignore")
-        )
-        return "\n- ".join(summary)
-
-    @staticmethod
-    def shieldsio_badge(value: float, label: str = "Total carbon emitted [gCO2eq]") -> str:
-        """Insert value in a Shields.io badge and return url
-
-        Insert total measurement value into a shields.io badge url
-
-        Parameters
-        ----------
-        value : float
-            The carbon footprint value in gCO2eq
-        label: str, default: 'Total carbon emitted'
-            The badge label to use
-
-        Returns
-        -------
-        str
-        """
-        payload = {
-            "style": "plastic",
-            "labelColor": "grey",
-        }
-        t = lambda t: urllib.parse.quote(t)
-        uri = "https://img.shields.io/badge/%s-%s-%s?" % (
-            t(label),
-            t("%0.2f gCO2eq" % value),
-            "black",
-        ) + urllib.parse.urlencode(payload)
-        return uri
-
-    @staticmethod
-    def shieldsio_endpoint(value: float, outfile: Path = None, label: str = "Total carbon emitted [gCO2eq]") -> Path:
-        """Insert value in a Shields.io json endpoint file
-
-        Parameters
-        ----------
-        value : float
-            The carbon footprint value in gCO2eq
-        outfile: :class:`pathlib.Path`, default: None
-            Path toward json file
-        label: str, default: 'Total carbon emitted'
-            The badge label to use
-
-        Returns
-        -------
-        :class:`pathlib.Path`
-        """
-
-        def save_to_json(label, message, outfile):
-            """Save a shields.io badge endpoint to a json file"""
-            data = {}
-            data["schemaVersion"] = 1
-            data["label"] = label
-            data["labelColor"] = "grey"
-            data["message"] = message
-            data["color"] = "black"
-            with open(outfile, "w") as f:
-                json.dump(data, f)
-            return outfile
-
-        return save_to_json(
-            label=label,
-            message="%0.2f gCO2eq" % value,
-            outfile=Path(outfile),
-        )
